Salvar back e front

+++ CurriculifyBack/curriculify/src/main/java/com/projeto/curriculify/ExpProfissional/ExpProfissionalRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExpProfissionalRepository extends JpaRepository<ExpProfissional, Long> {
+public interface ExpProfissionalRepository extends JpaRepository<ExpProfissional, Integer> {
 	List<ExpProfissional> findAll();
 	List<ExpProfissional> findAllByIdUsuario(Integer idUsuario);
 }
